cast as double

+++ src/Gui.java
@@ -35,7 +35,7 @@ public class Gui extends javax.swing.JFrame {
 	}
 	
 	public void train() {
-		GeneticAlgorithm ga = new GeneticAlgorithm(jSlider1.getValue()/100, jSlider2.getValue(), 169);
+		GeneticAlgorithm ga = new GeneticAlgorithm((double)jSlider1.getValue()/100, jSlider2.getValue(), 169);
 		try {
 			ga.start(jSlider3.getValue());
 		} catch (Exception e) {
